refactor: move rust package in root directory
There is no need to have the Rust package under `package` directory. The only consideration to do so
was to maintain a clean code structure, but the addition of package.json does not add much overhead
to the point where we'll have to keep this structure. By moving the Rust package out of the
`package` directory we also allow vscode-rls extension to work (which requires the Rust project to
be in the root).