feat: introduce a binary file for CLI integration
This commit does not implement any core CLI mechanism. It simply adds `main.rs` file that will allow
us to run the validate function from the terminal. This is not a production ready feature, the CLI
integration does not check for failures on passed arguments and the UX lacks documentation. This was
only added as a start point to fulfill project requirements and move to next tasks.