refactor: rename tokens module to json
This commit changes the `Tokens` struct name to `JSON` and all the tokens instances to
`json_document`. The reason for this change was the ambiguity resulted by the previous name. Tokens
struct is a general purpose struct that track the parsing of a JSON document and does not operate
only on JSON structural tokens.