feat: catch duplicate object names as errors
RFC 8259 does not strictly prohibit duplicate object names: "When the names within an object are not
unique, the behavior of software that receives such an object is unpredictable.". However, many
implementations choose to report duplicate names. We believe that prohibiting duplicate names are
crucial for most applications. For this reason we implement the check mechanism to report such
cases.